Desktop/FINAL-LESAVOT-Simple and complete/requirements.txt
PyQt5>=5.15.0
Pillow>=8.0.0
numpy>=1.19.0
scipy>=1.5.0
pyinstaller>=4.0
